IirisMaintenance as a separate component.
